docker/services/background-functions/codex-bgjobs/requirements.txt
azure-cosmosdb-table==1.0.6
azure-functions==1.5.0
azure-storage-blob==1.4.0
azure-storage-queue==12.1.4
eventlet==0.33.3
flasgger==0.9.5
flask==1.0.3
flask-migrate==2.4.0
flask-script==2.0.6
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.4.0
flask_cors==3.0.8
gmaps==0.9.0
ipywidgets==7.6.3
itsdangerous==2.0.1
Jinja2==3.0.3
kaleido==0.2.1
MarkupSafe==2.1.2
millify==0.1.1
numpy==1.24.2
openpyxl==3.0.6
pandas==1.5.3
parquet==1.3.1
plotly==5.8.0
psutil==5.8.0
# psycopg2==2.8.4
psycopg2-binary==2.9.5
pyarrow==7.0.0
pyjwt==2.4.0
pymongo==4.3.3
pyodbc==4.0.39
scikit-learn==1.2.1
scipy == 1.10.1
sendgrid==6.9.7
# sentry-sdk[flask]==1.24.0
simple-websocket
# snowflake-connector-python==2.7.6
SQLAlchemy==1.4.46
statsmodels == 0.13.5
Werkzeug==2.0.3
